Corrected Errored longitudes
Code was added to correct 83 records had incorrect longitudes when the data was viewed Tableau.
One Florida record that seems to have been incorrectly entered as 80.2 instead of -80.2, was negated to correct that, and
Other 82 records from Vermont and Florida were assigned a central longitude as per the respective states.
A new csv file was created to save the data with corrected longitudes that was then used in further analysis.